+++ packages/component-library/.storybook/preview.js
@@ -1,9 +1,14 @@
-import { Quasar } from "quasar";
+import { Quasar, Dark } from "quasar";
 import "@quasar/extras/material-icons/material-icons.css";
 import "quasar/dist/quasar.css";
 import { app } from "@storybook/vue3";
+import VChart from "vue-echarts";
 
-app.use(Quasar, { plugins: {} });
+import "echarts";
+import 'echarts/lib/chart/custom'
+
+app.use(Quasar, { plugins: {} }).component("v-chart", VChart);
+Dark.set(true);
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },