+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-rc.1](https://github.com/alercebroker/frontendcitos/compare/v1.1.0-rc.0...v1.1.0-rc.1) (2022-07-12)
+
+
+### Bug Fixes
+
+* using alt instead of onerror method ([be02be5](https://github.com/alercebroker/frontendcitos/commit/be02be5bbbcd72248356ffa28478501a1144d28e))
+* zoom corrected and select now changes the image ([dfc59bb](https://github.com/alercebroker/frontendcitos/commit/dfc59bbfdbd224d7d10fe77e823d8c8171a0d070))
+
+
+
+
+
+# [1.1.0-rc.0](https://github.com/alercebroker/frontendcitos/compare/v1.0.1-rc.16...v1.1.0-rc.0) (2022-07-12)
+
+
+### Features
+
+* dark mode params and onclick event ([34c830f](https://github.com/alercebroker/frontendcitos/commit/34c830fe52bc488a40439a8748935d1f3d0488e5))
+* LightCurvePlot completed ([c5a6b3f](https://github.com/alercebroker/frontendcitos/commit/c5a6b3fad613ad044be806f7de61a335c31268f8))
+* stamps with zoom and crosshair ([613d21d](https://github.com/alercebroker/frontendcitos/commit/613d21d437a7d032ec0f8c0ec9e52f6c4affdc0f))
+
+
+
+
+
 ## [1.0.1-rc.15](https://github.com/alercebroker/frontendcitos/compare/v1.0.1-rc.14...v1.0.1-rc.15) (2022-07-08)